Fixed reload bug
